@@ -1,6 +1,6 @@
 {
   "name": "tournado",
-  "version": "0.3.7",
+  "version": "0.3.8",
   "private": true,
   "sideEffects": false,
   "type": "module",
@@ -26,7 +26,7 @@
     "test:e2e:dev": "start-server-and-test dev http://localhost:3000 \"npx cypress open\"",
     "pretest:e2e:run": "npm run build",
     "reinstall": "rm -rf node_modules pnpm-lock.yaml && pnpm install",
-    "test:e2e:run": "node scripts/run-e2e-tests.cjs",
+    "test:e2e:run": "node scripts/run-e2e-tests.mjs",
     "typecheck": "tsc && tsc -p cypress",
     "postinstall": "husky",
     "prepare": "husky",
@@ -45,37 +45,37 @@
     "bcryptjs": "^3.0.2",
     "i18next": "^25.0.1",
     "i18next-browser-languagedetector": "^8.0.5",
-    "isbot": "^5.1.26",
-    "react": "18",
-    "react-dom": "18",
+    "isbot": "^5.1.27",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-i18next": "^15.4.1",
     "tiny-invariant": "^1.3.3"
   },
   "devDependencies": {
     "@faker-js/faker": "^9.7.0",
-    "@madrus/configs": "^1.6.1",
+    "@madrus/configs": "^3.0.3",
     "@remix-run/dev": "^2.16.5",
     "@tailwindcss/postcss": "^4.1.4",
     "@testing-library/cypress": "^10.0.3",
     "@testing-library/jest-dom": "^6.6.3",
     "@trivago/prettier-plugin-sort-imports": "^5.2.2",
     "@types/eslint": "^9.6.1",
-    "@types/node": "^22.14.1",
+    "@types/node": "^22.15.2",
     "@types/react": "18.3.20",
     "@types/react-dom": "18.3.6",
-    "@typescript-eslint/eslint-plugin": "^8.30.1",
-    "@typescript-eslint/parser": "^8.30.1",
+    "@typescript-eslint/eslint-plugin": "^8.31.0",
+    "@typescript-eslint/parser": "^8.31.0",
     "@vitejs/plugin-react": "^4.4.1",
-    "@vitest/coverage-v8": "^3.1.1",
+    "@vitest/coverage-v8": "^3.1.2",
     "commitizen": "^4.3.1",
     "cookie": "^1.0.2",
     "cross-env": "^7.0.3",
-    "cypress": "^14.3.1",
+    "cypress": "^14.3.2",
     "docsify-cli": "^4.4.4",
-    "eslint": "^9.25.0",
+    "eslint": "^9.25.1",
     "eslint-config-prettier": "^10.1.2",
-    "eslint-import-resolver-typescript": "^4.3.3",
-    "eslint-plugin-cypress": "^4.2.1",
+    "eslint-import-resolver-typescript": "^4.3.4",
+    "eslint-plugin-cypress": "^4.3.0",
     "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-jest": "^28.11.0",
     "eslint-plugin-jest-dom": "^5.5.0",
@@ -101,11 +101,11 @@
     "tsconfig-paths": "4.2.0",
     "tsx": "^4.19.3",
     "typescript": "^5.8.3",
-    "typescript-eslint": "^8.30.1",
-    "vite": "^6.3.2",
+    "typescript-eslint": "^8.31.0",
+    "vite": "^6.3.3",
     "vite-plugin-pwa": "^1.0.0",
     "vite-tsconfig-paths": "^5.1.4",
-    "vitest": "^3.1.1",
+    "vitest": "^3.1.2",
     "wait-on": "^8.0.3"
   },
   "engines": {
