@@ -1,11 +1,12 @@
+/* eslint-disable no-console */
 /**
  * This script runs the Cypress tests against the mock server.
  * It's used by Fly.io to run the tests in a production environment.
  * The reason to create is that the package.json script runs
  * with an error that the server is already stopped.
  */
-const { execSync, spawn } = require('child_process')
-const waitOn = require('wait-on')
+import { execSync, spawn } from 'child_process'
+import waitOn from 'wait-on'
 
 // Set environment variables
 process.env.PORT = '8811'
