@@ -2,9 +2,9 @@
 import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { updateServiceWorker } from '@/utils/serviceWorker'
+import { updateServiceWorker } from '@/utils/serviceWorkerRegistration'
 
-export function UpdatePrompt() {
+export function UpdatePrompt(): JSX.Element | null {
   const { t } = useTranslation()
   const [isClient, setIsClient] = useState(false)
   const [showUpdatePrompt, setShowUpdatePrompt] = useState(false)
