@@ -5,20 +5,28 @@ import type { User } from '@prisma/client'
 
 export type { User } from '@prisma/client'
 
-export async function getUserById(id: User['id']) {
-  return prisma.user.findUnique({ where: { id } })
-}
+export const getUserById = async (id: User['id']): Promise<User | null> =>
+  prisma.user.findUnique({ where: { id } })
 
-export async function getUserByEmail(email: User['email']) {
-  return prisma.user.findUnique({ where: { email } })
-}
+export const getUserByEmail = async (email: User['email']): Promise<User | null> =>
+  prisma.user.findUnique({ where: { email } })
 
-export async function createUser(email: User['email'], password: string) {
+export async function createUser(
+  email: User['email'],
+  password: string,
+  firstName: User['firstName'],
+  lastName: User['lastName'],
+  role: User['role']
+): Promise<User> {
   const hashedPassword = await bcrypt.hash(password, 10)
 
   return prisma.user.create({
+    // eslint-disable-next-line id-blacklist
     data: {
       email,
+      firstName,
+      lastName,
+      role,
       password: {
         create: {
           hash: hashedPassword,
@@ -28,11 +36,13 @@ export async function createUser(email: User['email'], password: string) {
   })
 }
 
-export async function deleteUserByEmail(email: User['email']) {
-  return prisma.user.delete({ where: { email } })
-}
+export const deleteUserByEmail = async (email: User['email']): Promise<User> =>
+  prisma.user.delete({ where: { email } })
 
-export async function verifyLogin(email: User['email'], password: string) {
+export async function verifyLogin(
+  email: User['email'],
+  password: string
+): Promise<Omit<User & { password: { hash: string } }, 'password'> | null> {
   const userWithPassword = await prisma.user.findUnique({
     where: { email },
     include: {
