@@ -1,5 +1,4 @@
 import type { LoaderFunctionArgs } from '@remix-run/node'
 
-export const loader = async ({ request }: LoaderFunctionArgs) => {
-  return new Response(null, { status: 204 })
-}
+export const loader = async ({ request: _ }: LoaderFunctionArgs): Promise<Response> =>
+  new Response(null, { status: 204 })
