@@ -1,9 +1,10 @@
+/* eslint-disable no-console */
 // learn more: https://fly.io/docs/reference/configuration/#services-http_checks
 import type { LoaderFunctionArgs } from '@remix-run/node'
 
 import { prisma } from '@/db.server'
 
-export const loader = async ({ request }: LoaderFunctionArgs) => {
+export const loader = async ({ request }: LoaderFunctionArgs): Promise<Response> => {
   const host = request.headers.get('X-Forwarded-Host') ?? request.headers.get('host')
 
   try {
