@@ -3,6 +3,7 @@ import { redirect } from '@remix-run/node'
 
 import { logout } from '@/utils/session.server'
 
-export const action = async ({ request }: ActionFunctionArgs) => logout(request)
+export const action = async ({ request }: ActionFunctionArgs): Promise<Response> =>
+  logout(request)
 
-export const loader = async () => redirect('/')
+export const loader = async (): Promise<Response> => redirect('/')
