@@ -1,6 +1,6 @@
 import { type ReactNode, useEffect, useState } from 'react'
 
-export function ClientOnly({ children }: { children: ReactNode }) {
+export function ClientOnly({ children }: { children: ReactNode }): ReactNode {
   const [mounted, setMounted] = useState(false)
 
   useEffect(() => {
